+++ src/simpleGit.ts
@@ -1,7 +1,7 @@
 import { spawnSync } from "child_process";
-import { FileSystemAdapter } from "obsidian";
+import { FileSystemAdapter, normalizePath, Notice } from "obsidian";
 import * as path from "path";
-import { sep } from "path";
+import { sep, } from "path";
 import * as simple from "simple-git";
 import simpleGit, { DefaultLogFields } from "simple-git";
 import { GitManager } from "./gitManager";
@@ -12,23 +12,26 @@ export class SimpleGit extends GitManager {
     git: simple.SimpleGit;
     constructor(plugin: ObsidianGit) {
         super(plugin);
-
-        this.setGitInstance();
-
     }
 
-    private async setGitInstance() {
+    async setGitInstance(ignoreError: boolean = false): Promise<void> {
         if (this.isGitInstalled()) {
             const adapter = this.app.vault.adapter as FileSystemAdapter;
             const path = adapter.getBasePath();
-            let extraPath = "";
+            let basePath = path;
             // Because the basePath setting is a relative path, a leading `/` must
             // be appended before concatenating with the path.
             if (this.plugin.settings.basePath) {
-                extraPath = sep + this.plugin.settings.basePath;
+                const exists = await adapter.exists(normalizePath(this.plugin.settings.basePath));
+                if (exists) {
+                    basePath = path + sep + this.plugin.settings.basePath;
+                } else if (!ignoreError) {
+                    new Notice("ObsidianGit: Base path does not exist");
+                }
             }
+
             this.git = simpleGit({
-                baseDir: path + extraPath,
+                baseDir: basePath,
                 binary: this.plugin.settings.gitPath || undefined,
                 config: ["core.quotepath=off"]
             });
@@ -43,29 +46,48 @@ export class SimpleGit extends GitManager {
     }> {
         this.plugin.setState(PluginState.status);
         const status = await this.git.status((err) => this.onError(err));
-
         this.plugin.setState(PluginState.idle);
         return {
             changed: status.files.filter((e) => e.working_dir !== " ").map((e) => {
-                const res = this.formatPath(e.path);
-                e.path = res.path;
-                e.from = res.from;
-                e.working_dir = e.working_dir === "?" ? "U" : e.working_dir;
-                return e;
+                const res = this.formatPath(e);
+
+                return <FileStatusResult>{
+                    path: res.path,
+                    from: res.from,
+                    working_dir: e.working_dir === "?" ? "U" : e.working_dir,
+                    vault_path: this.getVaultPath(res.path),
+                };
             }),
             staged: status.files.filter((e) => e.index !== " " && e.index != "?").map((e) => {
-                const res = this.formatPath(e.path, e.index === "R");
-                e.path = res.path;
-                e.from = res.from;
-                return e;
+                const res = this.formatPath(e, e.index === "R");
+                return <FileStatusResult>{
+                    path: res.path,
+                    from: res.from,
+                    index: e.index,
+                    vault_path: this.getVaultPath(res.path),
+                };
             }),
-            conflicted: status.conflicted.map((e) => this.formatPath(e).path),
+            conflicted: status.conflicted.map((e) => this.formatPath({
+                path: e,
+                from: undefined,
+                index: undefined,
+                working_dir: undefined
+            }).path),
         };
     }
 
+    getVaultPath(path: string): String {
+        if (this.plugin.settings.basePath) {
+            return this.plugin.settings.basePath + "/" + path;
+        } else {
+            return path;
+        }
+    }
+
     //Remove wrong `"` like "My file.md"
-    formatPath(path: string, renamed: boolean = false): { path: string, from?: string; } {
-        function format(path: string): string {
+    formatPath(path: simple.FileStatusResult, renamed: boolean = false): { path: string, from?: string; } {
+        function format(path?: string): string {
+            if (path == undefined) return undefined;
 
             if (path.startsWith('"') && path.endsWith('"')) {
                 return path.substring(1, path.length - 1);
@@ -74,15 +96,13 @@ export class SimpleGit extends GitManager {
             }
         }
         if (renamed) {
-            const paths = path.split(" -> ").map((e) => format(e));
-
             return {
-                from: paths[0],
-                path: paths[1],
+                from: format(path.from),
+                path: format(path.path),
             };
         } else {
             return {
-                path: format(path)
+                path: format(path.path)
             };
         }
     }
@@ -99,7 +119,7 @@ export class SimpleGit extends GitManager {
                     if (!(args.contains("submodule") && args.contains("foreach"))) return;
 
                     let body = "";
-                    let root = (this.app.vault.adapter as FileSystemAdapter).getBasePath() + (this.plugin.settings.basePath ? sep + this.plugin.settings.basePath : "");
+                    let root = (this.app.vault.adapter as FileSystemAdapter).getBasePath() + (this.plugin.settings.basePath ? "/" + this.plugin.settings.basePath : "");
                     stdout.on('data', (chunk) => {
                         body += chunk.toString('utf8');
                     });
@@ -110,7 +130,7 @@ export class SimpleGit extends GitManager {
                         submods = submods.map(i => {
                             let submod = i.match(/'([^']*)'/);
                             if (submod != undefined) {
-                                return root + sep + submod[1] + sep;
+                                return root + "/" + submod[1] + sep;
                             }
                         });
 
@@ -193,7 +213,7 @@ export class SimpleGit extends GitManager {
             await this.git.subModule(["update", "--remote", "--merge", "--recursive"], (err) => this.onError(err));
 
         const branchInfo = await this.branchInfo();
-        const localCommit = await this.git.revparse([branchInfo.tracking], (err) => this.onError(err));
+        const localCommit = await this.git.revparse([branchInfo.current], (err) => this.onError(err));
 
         await this.git.fetch((err) => this.onError(err));
         const upstreamCommit = await this.git.revparse([branchInfo.tracking], (err) => this.onError(err));
@@ -211,10 +231,6 @@ export class SimpleGit extends GitManager {
                     }
                 } catch (err) {
                     this.plugin.displayError(`Pull failed (${this.plugin.settings.syncMethod}): ${err.message}`);
-                    const status = await this.status();
-                    if (status.conflicted.length > 0) {
-                        this.plugin.handleConflict(status.conflicted);
-                    }
                     return;
                 }
 
@@ -226,8 +242,9 @@ export class SimpleGit extends GitManager {
                     this.plugin.displayError(`Sync failed (${this.plugin.settings.syncMethod}): ${err.message}`);
                 }
             }
+            const afterMergeCommit = await this.git.revparse([branchInfo.current], (err) => this.onError(err));
 
-            const filesChanged = await this.git.diff([`${localCommit}..${upstreamCommit}`, '--name-only']);
+            const filesChanged = await this.git.diff([`${localCommit}..${afterMergeCommit}`, '--name-only']);
             return filesChanged.split(/\r\n|\r|\n/).filter((value) => value.length > 0).length;
         } else {
             return 0;
@@ -239,7 +256,7 @@ export class SimpleGit extends GitManager {
         const status = await this.git.status();
         const trackingBranch = status.tracking;
         const currentBranch = status.current;
-        const remoteChangedFiles = (await this.git.diffSummary([currentBranch, trackingBranch])).changed;
+        const remoteChangedFiles = (await this.git.diffSummary([currentBranch, trackingBranch], (err) => this.onError(err))).changed;
 
         this.plugin.setState(PluginState.push);
         if (this.plugin.settings.updateSubmodules) {
@@ -286,13 +303,17 @@ export class SimpleGit extends GitManager {
         };
     }
 
-    async log(file?: string): Promise<ReadonlyArray<DefaultLogFields>> {
-        const res = await this.git.log({ file: file, }, (err) => this.onError(err));
+    async log(file?: string, relativeToRepo: boolean = false): Promise<ReadonlyArray<DefaultLogFields>> {
+        const path = (relativeToRepo && this.plugin.settings.basePath) ? file : file?.substring(this.plugin.settings.basePath.length + 1);
+
+        const res = await this.git.log({ file: path, }, (err) => this.onError(err));
         return res.all;
     }
 
-    async show(commitHash: string, file: string): Promise<string> {
-        return this.git.show([commitHash + ":" + file], (err) => this.onError(err));
+    async show(commitHash: string, file: string, relativeToRepo: boolean = false): Promise<string> {
+        const path = (relativeToRepo && this.plugin.settings.basePath) ? file : file.substring(this.plugin.settings.basePath.length + 1);
+
+        return this.git.show([commitHash + ":" + path], (err) => this.onError(err));
     }
 
     async checkout(branch: string): Promise<void> {
@@ -361,7 +382,7 @@ export class SimpleGit extends GitManager {
     }
 
     updateBasePath(basePath: string) {
-        this.setGitInstance();
+        this.setGitInstance(true);
     }
 
     async getDiffString(filePath: string, stagedChanges = false): Promise<string> {